+++ Index.php
@@ -4,5 +4,5 @@ require_once("MysqlUtil.php");
 // create an object that will serve as our database.
 $db = new Database; 
 require_once("views/Header.php");
-    require_once("ParentForum.php");
+require_once("ParentForum.php");
 require_once("views/Footer.php");
